Implement Generic Services in Controllers

+++ src/main/java/com/anderson/tiendaback/controllers/CategoryController.java
@@ -1,14 +1,15 @@
 package com.anderson.tiendaback.controllers;
 
+import java.net.URI;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import com.anderson.tiendaback.exception.ModelNotFoundException;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
+
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,15 +17,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.anderson.tiendaback.exception.ResourceNotFoundException;
+import static org.springframework.http.HttpStatus.NO_CONTENT;
+import static org.springframework.http.HttpStatus.OK;
+
 import com.anderson.tiendaback.models.Category;
 import com.anderson.tiendaback.services.CategoryService;
 import com.anderson.tiendaback.dto.CategoryDTO;
 
-import jakarta.validation.ValidationException;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 @RestController
 @RequestMapping("/api/categories")
@@ -34,48 +36,47 @@ public class CategoryController {
 	private CategoryService categoryService;
 	
 	@Autowired
-	private Mapper mapper;
-	
-	@GetMapping(params = {"page","size","sortDir","sort"})
-	public ResponseEntity<List<Category>> getCategory(@RequestParam("page")int page,@RequestParam("size")int size,@RequestParam("sortDir")String sortDir,@RequestParam("sort")String sort){
-		List<Category> categorys = categoryService.getAll(page, size, sortDir, sort).stream().map(category->{
-			return category;
-		}).collect(Collectors.toList());
-		return new ResponseEntity<List<Category>>(categorys,HttpStatus.OK);
+	private ModelMapper mapper;
+
+	@GetMapping
+	public ResponseEntity<List<CategoryDTO>> findAll() {
+		List<CategoryDTO> list = categoryService.findAll().stream().map(p -> mapper.map(p, CategoryDTO.class)).collect(Collectors.toList());
+
+		return new ResponseEntity<>(list, OK);
 	}
-	
+
 	@GetMapping("/{id}")
-	public ResponseEntity<Category>getCategoryById(@PathVariable UUID id){
-		Optional<Category> category = categoryService.getOne(id);
-		if(!category.isPresent()) {
-			new ResourceNotFoundException("Id "+id+ "is not existed");
+	public ResponseEntity<CategoryDTO> findById(@PathVariable("id") UUID id) {
+		Category obj = categoryService.findById(id);
+		if (obj == null) {
+			throw new ModelNotFoundException("ID NOT FOUND: " + id);
+		} else {
+			return new ResponseEntity<>(mapper.map(obj, CategoryDTO.class), OK);
 		}
-		return ResponseEntity.ok(category.get());
 	}
-	
+
+
 	@PostMapping
-	public ResponseEntity<Category>createCategory(@RequestBody CategoryDTO categoryViewModel, BindingResult bindingResult){
-		if(bindingResult.hasErrors()) {
-			throw new ValidationException();
-		}
-		Category category = this.mapper.convertToCategoryEntity(categoryViewModel);
-		return new ResponseEntity<>(categoryService.insertOrUpdate(category),HttpStatus.CREATED);
+	public ResponseEntity<Void> save(@RequestBody CategoryDTO dto){
+		Category obj = categoryService.save(mapper.map(dto, Category.class));
+		//localhost:8080/category/5
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(obj.getCategoryId()).toUri();
+		return ResponseEntity.created(location).build();
 	}
-	
-	@PutMapping("/{id}")
-	public ResponseEntity<Category>updateCategory(@PathVariable UUID id,@RequestBody CategoryDTO categoryViewModel){
-		if(!categoryService.getOne(id).isPresent()) {
-        	new ResourceNotFoundException("Category not found with id " + id);
-		}
-		Category category = this.mapper.convertToCategoryEntity(categoryViewModel);
-		return  ResponseEntity.ok(categoryService.insertOrUpdate(category));
+
+	@PutMapping
+	public ResponseEntity<Category> update(@RequestBody CategoryDTO dto) {
+		Category obj = categoryService.update(mapper.map(dto, Category.class));
+		return new ResponseEntity<>(obj, OK);
 	}
-	
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?>deleteCategory(@PathVariable UUID id){
-		return categoryService.getOne(id).map(fabric -> {
-			categoryService.delete(id);
-			return ResponseEntity.ok().build();
-		}).orElseThrow(() -> new ResourceNotFoundException("Category not found with id " + id));
+	public ResponseEntity<Void> delete(@PathVariable("id") UUID id) {
+		Category obj = categoryService.findById(id);
+		if (obj == null) {
+			throw new ModelNotFoundException("ID NOT FOUND: " + id);
+		}
+		categoryService.delete(id);
+		return new ResponseEntity<>(NO_CONTENT);
 	}
 }
