+++ src/main/java/com/anderson/tiendaback/controllers/ProductController.java
@@ -1,15 +1,18 @@
 package com.anderson.tiendaback.controllers;
 
+import java.net.URI;
 import java.util.List;
-import java.util.Optional;
+
 import java.util.UUID;
 import java.util.stream.Collectors;
 
 
+import com.anderson.tiendaback.dto.ProductDTO;
+import com.anderson.tiendaback.exception.ModelNotFoundException;
+import com.anderson.tiendaback.models.Product;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,16 +20,15 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 
-import com.anderson.tiendaback.exception.ResourceNotFoundException;
-import com.anderson.tiendaback.models.Product;
 import com.anderson.tiendaback.services.ProductService;
-import com.anderson.tiendaback.dto.ProductDTO;
 
-import jakarta.validation.ValidationException;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+import static org.springframework.http.HttpStatus.NO_CONTENT;
+import static org.springframework.http.HttpStatus.OK;
 
 @RestController
 @RequestMapping("/api/products")
@@ -36,49 +38,48 @@ public class ProductController {
 	private ProductService productService;
 	
 	@Autowired
-	private Mapper mapper;
-	
-	@GetMapping(params = {"page","size","sortDir","sort"})
-	public ResponseEntity<List<Product>> getProduct(@RequestParam("page")int page,@RequestParam("size")int size,@RequestParam("sortDir")String sortDir,@RequestParam("sort")String sort){
-		List<Product> products = productService.getAll(page, size, sortDir, sort).stream().map(product->{
-			return product;
-		}).collect(Collectors.toList());
-		return new ResponseEntity<List<Product>>(products,HttpStatus.OK);
+	private ModelMapper mapper;
+
+	@GetMapping
+	public ResponseEntity<List<ProductDTO>> findAll() {
+		List<ProductDTO> list = productService.findAll().stream().map(p -> mapper.map(p, ProductDTO.class)).collect(Collectors.toList());
+
+		return new ResponseEntity<>(list, OK);
 	}
-	
+
 	@GetMapping("/{id}")
-	public ResponseEntity<Product>getProductById(@PathVariable UUID id){
-		Optional<Product> product = productService.getOne(id);
-		if(!product.isPresent()) {
-			new ResourceNotFoundException("Id "+id+ "is not existed");
+	public ResponseEntity<ProductDTO> findById(@PathVariable("id") UUID id) {
+		Product obj = productService.findById(id);
+		if (obj == null) {
+			throw new ModelNotFoundException("ID NOT FOUND: " + id);
+		} else {
+			return new ResponseEntity<>(mapper.map(obj, ProductDTO.class), OK);
 		}
-		return ResponseEntity.ok(product.get());
 	}
-	
+
+
 	@PostMapping
-	public ResponseEntity<Product>createProduct(@RequestBody ProductDTO productViewModel, BindingResult bindingResult){
-		if(bindingResult.hasErrors()) {
-			throw new ValidationException();
-		}
-		Product product = this.mapper.convertToProductEntity(productViewModel);
-		return new ResponseEntity<>(productService.insertOrUpdate(product),HttpStatus.CREATED);
+	public ResponseEntity<Void> save(@RequestBody ProductDTO dto){
+		Product obj = productService.save(mapper.map(dto, Product.class));
+		//localhost:8080/product/5
+		URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(obj.getProductId()).toUri();
+		return ResponseEntity.created(location).build();
 	}
-	
-	@PutMapping("/{id}")
-	public ResponseEntity<Product>updateProduct(@PathVariable UUID id,@RequestBody ProductDTO productViewModel){
-		if(!productService.getOne(id).isPresent()) {
-        	new ResourceNotFoundException("Product not found with id " + id);
-		}
-		Product product = this.mapper.convertToProductEntity(productViewModel);
-		return  ResponseEntity.ok(productService.insertOrUpdate(product));
+
+	@PutMapping
+	public ResponseEntity<Product> update(@RequestBody ProductDTO dto) {
+		Product obj = productService.update(mapper.map(dto, Product.class));
+		return new ResponseEntity<>(obj, OK);
 	}
-	
+
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?>deleteProduct(@PathVariable UUID id){
-		return productService.getOne(id).map(fabric -> {
-			productService.delete(id);
-			return ResponseEntity.ok().build();
-		}).orElseThrow(() -> new ResourceNotFoundException("Product not found with id " + id));
+	public ResponseEntity<Void> delete(@PathVariable("id") UUID id) {
+		Product obj = productService.findById(id);
+		if (obj == null) {
+			throw new ModelNotFoundException("ID NOT FOUND: " + id);
+		}
+		productService.delete(id);
+		return new ResponseEntity<>(NO_CONTENT);
 	}
